server_backup: support running in fly.io
This updates the question generator server to support reading
environment variables from process.env rather than the `.env` file.
This is necessary to allow fly.io read secrets from the environment so
we don't have to publish them in github.

It only does this when DEPLOYMENT=production is set in the environment.
Otherwise, it reads them from the `.env` file so local development still
works as expected.

It also adds a `fly.toml` file to deploy this to fly.io, including a
`DEPLOYMENT=production` environment variable.